requirements.txt
Django
djangorestframework
djangorestframework_simplejwt
django-cors-headers
psycopg2-binary
python-decouple
celery
redis